requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
appier==1.34.6
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dateparser==1.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
fuzzywuzzy==0.18.0
git-filter-repo==2.47.0
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.1
MarkupSafe==3.0.2
multidict==6.4.4
ngrok==1.4.0
packaging==25.0
propcache==0.3.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.41
twilio==9.6.1
twilio_api==0.1.1
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
